+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActFramework Change Log
 
 **1.8.8**
+* `CacheFor` - make it allow skip `Cache-Control` header #837
+* `ResponseCache` - content disposition shall be cached #838
 * `CacheFor` key shall include `Accept` type #836
 * Add render image support #834
 * Optimise `LogSupport` for `DestroyableBase` #833